Unifying the "series executor" usage + Moving it (and sibling helpers) to a separate module.
