+++ core-rust-bridge/src/main/java/com/radixdlt/statecomputer/commit/PrepareRequest.java
@@ -64,12 +64,10 @@
 
 package com.radixdlt.statecomputer.commit;
 
-import com.radixdlt.rev2.ComponentAddress;
 import com.radixdlt.sbor.codec.CodecMap;
 import com.radixdlt.sbor.codec.StructCodec;
 import com.radixdlt.transactions.RawLedgerTransaction;
 import com.radixdlt.transactions.RawNotarizedTransaction;
-import com.radixdlt.utils.UInt64;
 import java.util.List;
 
 public record PrepareRequest(
@@ -77,12 +75,7 @@ public record PrepareRequest(
     List<RawLedgerTransaction> ancestorTransactions,
     LedgerHashes ancestorLedgerHashes,
     List<RawNotarizedTransaction> proposedTransactions,
-    boolean isFallback,
-    UInt64 epoch,
-    UInt64 roundNumber,
-    List<ComponentAddress> gapRoundLeaderAddresses,
-    ComponentAddress proposerAddress,
-    long proposerTimestampMs) {
+    RoundHistory roundHistory) {
   public static void registerCodec(CodecMap codecMap) {
     codecMap.register(
         PrepareRequest.class,
